pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "c3po_lrs_experiments"
version = "0.1.0"
dependencies = [
  "hydra-core>=1.3",
  "wandb>=0.16",
  "torch>=2.1",
  "transformers>=4.38",
  "datasets>=2.15",
  "peft>=0.7",
  "bitsandbytes>=0.43.1",
  "optuna>=3.5",
  "pynvml>=11.5",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "pandas>=2.1",
  "pyyaml>=6.0",
  "scikit-learn>=1.3",
  "scipy>=1.11",
]